🎨 fix todos

+++ test/fuzz/asset-modules/UniswapV3AssetModule/processDirectDeposit.fuzz.t.sol
@@ -272,8 +272,8 @@ contract ProcessDirectDeposit_UniswapV3AssetModule_Fuzz_Test is UniswapV3AssetMo
             vm.assume(amount1 + initialExposure1 < maxExposure1);
 
             // And: Usd value of underlying assets does not overflow.
-            vm.assume(amount0 + initialExposure0 <= type(uint256).max / priceToken0 / 10 ** (18 - 0)); // divided by 10 ** (18 - DecimalsOracle).
-            vm.assume(amount1 + initialExposure1 <= type(uint256).max / priceToken0 / 10 ** (18 - 0)); // divided by 10 ** (18 - DecimalsOracle).
+            vm.assume(amount0 + initialExposure0 <= type(uint112).max / priceToken0 / 10 ** (18 - 0)); // divided by 10 ** (18 - DecimalsOracle).
+            vm.assume(amount1 + initialExposure1 <= type(uint112).max / priceToken0 / 10 ** (18 - 0)); // divided by 10 ** (18 - DecimalsOracle).
         }
 
         // Add underlying tokens and its oracles to Arcadia.
